+++ src/main/java/com/bookbla/americano/domain/admin/controller/AdminMemberController.java
@@ -1,8 +1,5 @@
 package com.bookbla.americano.domain.admin.controller;
 
-import com.bookbla.americano.domain.admin.controller.dto.request.AdminMemberKakaoRoomStatusUpdateRequest;
-import com.bookbla.americano.domain.admin.controller.dto.request.AdminMemberProfileImageStatusUpdateRequest;
-import com.bookbla.americano.domain.admin.controller.dto.request.AdminMemberProfileStudentIdStatusUpdateRequest;
 import com.bookbla.americano.domain.admin.controller.dto.response.AdminMemberKakaoRoomResponses;
 import com.bookbla.americano.domain.admin.controller.dto.response.AdminMemberProfileImageResponses;
 import com.bookbla.americano.domain.admin.controller.dto.response.AdminMemberProfileStatusResponse;
@@ -10,15 +7,11 @@ import com.bookbla.americano.domain.admin.controller.dto.response.AdminMemberRea
 import com.bookbla.americano.domain.admin.controller.dto.response.AdminMemberStudentIdResponses;
 import com.bookbla.americano.domain.admin.service.AdminMemberService;
 import io.swagger.v3.oas.annotations.Operation;
-import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -60,32 +53,12 @@ public class AdminMemberController {
         return ResponseEntity.ok(adminMemberKakaoRoomResponses);
     }
 
-    @Operation(summary = "카카오톡 오픈 채팅방 상태 변경 API")
-    @PatchMapping("/member-verifies/{memberVerifyId}/pending/kakao")
-    public ResponseEntity<Void> updateKakaoRoomUrlPendingMemberStatus(
-            @PathVariable Long memberVerifyId,
-            @RequestBody @Valid AdminMemberKakaoRoomStatusUpdateRequest request
-    ) {
-        adminMemberService.updateMemberKakaoRoomStatus(request.toDto(memberVerifyId));
-        return ResponseEntity.ok().build();
-    }
-
     @Operation(summary = "프로필 이미지 승인 대기 회원 조회 API")
     @GetMapping("/members/pending/profile-image")
     public ResponseEntity<AdminMemberProfileImageResponses> readProfileImagePendingMembers(Pageable pageable) {
         PageRequest pageRequest = PageRequest.of(pageable.getPageNumber() - 1, pageable.getPageSize(), pageable.getSort());
         AdminMemberProfileImageResponses adminMemberProfileImageResponses = adminMemberService.readProfileImagePendingMembers(pageRequest);
         return ResponseEntity.ok(adminMemberProfileImageResponses);
-    }
-
-    @Operation(summary = "프로필 이미지 상태 변경 API")
-    @PatchMapping("/member-verifies/{memberVerifyId}/pending/profile-image")
-    public ResponseEntity<Void> updateProfileImagePendingMemberStatus(
-            @PathVariable Long memberVerifyId,
-            @RequestBody @Valid AdminMemberProfileImageStatusUpdateRequest request
-    ) {
-        adminMemberService.updateMemberImageStatus(request.toDto(memberVerifyId));
-        return ResponseEntity.ok().build();
     }
 
     @Operation(summary = "학생증 승인 대기 회원 조회 API")
@@ -96,13 +69,4 @@ public class AdminMemberController {
         return ResponseEntity.ok(adminMemberStudentIdResponses);
     }
 
-    @Operation(summary = "학생증 이미지 상태 변경 API")
-    @PatchMapping("/member-verifies/{memberVerifyId}/pending/student-id/image")
-    public ResponseEntity<Void> updateStudentIdPendingMemberStatus(
-            @PathVariable Long memberVerifyId,
-            @RequestBody @Valid AdminMemberProfileStudentIdStatusUpdateRequest request
-    ) {
-        adminMemberService.updateMemberStudentIdStatus(request.toDto(memberVerifyId));
-        return ResponseEntity.ok().build();
-    }
 }
